server/requirements.txt
fastapi
uvicorn
pydantic
openai
langchain
pinecone
python-dotenv
tiktoken
langchain_community
langchain_experimental
langchain-huggingface
langchain_pinecone
cryptography
pypdf
python-multipart
streamlit
langchain_openai